refactor: update types and move search client to constants
