+++ src/utils/contexts/search.tsx
@@ -1,6 +1,6 @@
 import { createContext, Dispatch, SetStateAction, useState } from 'react'
 import type { DocumentationTitle, UpdatesTitle } from 'utils/typings/unionTypes'
-import type { SearchDataItemProps } from 'components/search-results'
+import { Hit } from 'react-instantsearch-core'
 
 type FilterType = DocumentationTitle | UpdatesTitle | ''
 
@@ -8,7 +8,7 @@ type SearchContextType = {
   filterSelectedSection: FilterType
   changeFilterSelectedSection: Dispatch<SetStateAction<FilterType>>
   ocurrenceCount: Map<FilterType, number>
-  updateOcurrenceCount: (resultsData: SearchDataItemProps[]) => void
+  updateOcurrenceCount: (resultsData: Hit[]) => void
 }
 
 export const SearchContext = createContext<SearchContextType>({
@@ -25,12 +25,12 @@ const SearchContextProvider: React.FC = ({ children }) => {
     Map<FilterType, number>
   >(new Map())
 
-  const updateOcurrenceCount = (resultsData: SearchDataItemProps[]) => {
+  const updateOcurrenceCount = (resultsData: Hit[]) => {
     const ocurrenceCountMap = new Map<FilterType, number>()
     resultsData.forEach((result) => {
       ocurrenceCountMap.set(
-        result.doc,
-        ocurrenceCountMap.get(result.doc)! + 1 || 1
+        result.docType as FilterType,
+        (ocurrenceCountMap.get(result.docType as FilterType) || 0) + 1
       )
     })
     ocurrenceCountMap.set('', resultsData.length)
